Change CI image
The AWX EE image contains a cache folder in the user's home which
is owned by root and this breaks molecule. So instead a python:3 image
is used which should work just as fine even though it diverge a little bit
from what is actually used in AAP / AWX when executing the role that
way.